backend/adapter/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.6
pyjwt==2.8.0
redis==5.0.1
rq==1.15.1
prometheus-client==0.19.0
neo4j==5.16.0
# graphiti-core # If available, otherwise we use REST
openai==1.10.0
pytest==8.0.0
pytest-asyncio==0.23.4
